notifications ui

+++ app/controllers/contacts_controller.rb
@@ -16,16 +16,20 @@ class ContactsController < ApplicationController
 
   def destroy
     contact_user = Contact.where( :user_id => current_user.id, :person_id => params[ :person_id ] ).first
-    if !contact_user.mutual?
-      contact_user.destroy
-    else
-      contact_user.update_attributes(:receiving => false)
+    if contact_user
+      if !contact_user.mutual?
+        contact_user.destroy
+      else
+        contact_user.update_attributes(:receiving => false)
+      end
     end 
     contact_person = Contact.where( :user_id =>params[ :person_id ] , :person_id => current_user.id).first
-    if !contact_person.mutual?
-      contact_person.destroy
-    else
-      contact_person.update_attributes(:sharing => false)
+    if contact_person
+      if !contact_person.mutual?
+        contact_person.destroy
+      else
+        contact_person.update_attributes(:sharing => false)
+      end
     end 
     render :nothing=>true,:status => 200
   end
